The code didn't decrypt
I want to work on this some more because every time I tried to run it it gave me this error

File "/home/liveuser/ransomware/decrypt.py", line 27, in <module>
    contents_decrypted = Fernet(secretkey).decrypt(contents)
  File "/usr/lib64/python3.10/site-packages/cryptography/fernet.py", line 34, in __init__
    raise ValueError(
ValueError: Fernet key must be 32 url-safe base64-encoded bytes.

But it did work for NetworkChuck, and Stack overflow says its because its reading the key as plain text insted of bytes. I don't know how to fix it. For now I'll leave it and try running it after turning my VM on/off.